chore: create thread-component to avoid to become thread list into a client component
